Problem: integer overflow when calculating sizes (may happen e.g. with very large images) (#1148)
* Problem: integer overflow when calculating sizes (may happen e.g. with very large images)
Solution: change some types from (unsigned) long to size_t

# Conflicts:
#	dlib/dnn/tensor.h

* Fix the fact that std::numeric_limits<unsigned long>::max() isn't always the same number

* Revert serialization changes

* Review fix: use long long instead of size_t

* From long to long long all the way

* Change more types to (hopefully) make the compiler happy

* Change many more types to size_t

* Change even more types to size_t

* Minor type changes

+++ dlib/array2d/array2d_kernel.h
@@ -312,8 +312,8 @@ namespace dlib
             }
         }
 
-        unsigned long size (
-        ) const { return static_cast<unsigned long>(nc_ * nr_); }
+        size_t size (
+        ) const { return static_cast<size_t>(nc_) * static_cast<size_t>(nr_); }
 
         long width_step (
         ) const
