+++ dlib/gui_widgets/base_widgets_abstract.h
@@ -967,7 +967,7 @@ namespace dlib
                       widgets in this group and the upper left corner of get_rect(). 
         !*/
 
-        unsigned long size (
+        size_t size (
         ) const;
         /*!
             ensures
@@ -1561,7 +1561,7 @@ namespace dlib
                 - the menu_item in this with the index idx has been disabled
         !*/
 
-        unsigned long size (
+        size_t size (
         ) const;
         /*!
             ensures
