+++ dlib/sequence/sequence_kernel_2.h
@@ -103,7 +103,7 @@ namespace dlib
             );
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             bool at_start (
@@ -430,7 +430,7 @@ namespace dlib
         typename T,
         typename mem_manager
         >
-    unsigned long sequence_kernel_2<T,mem_manager>::
+    size_t sequence_kernel_2<T,mem_manager>::
     size (
     ) const
     {
